correct message

+++ src/main/java/com/mary/sharik/model/entity/Product.java
@@ -2,6 +2,8 @@ package com.mary.sharik.model.entity;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -9,7 +11,7 @@ public class Product {
     @Id
     private String id;
     private String name;
-    private Integer price;
+    private BigDecimal price;
     private Integer amountLeft;
     private String description;
     private String imageUrl;
